refactor(mf2): Rename src/data-model/ files, drop getMessageValue export
