+++ packages/mf2-messageformat/src/data-model/resolve-literal.ts
@@ -1,6 +1,6 @@
-import type { Context } from '../../format-context.js';
-import { MessageFunctionContext } from '../../runtime/index.js';
-import type { Literal } from '../types.js';
+import type { Context } from '../format-context.js';
+import { MessageFunctionContext } from '../runtime/index.js';
+import type { Literal } from './types.js';
 
 export function resolveLiteral(ctx: Context, lit: Literal) {
   const msgCtx = new MessageFunctionContext(ctx, `|${lit.value}|`);
