+++ packages/mf2-messageformat/src/data-model/resolve-value.ts
@@ -1,6 +1,6 @@
-import type { Context } from '../../format-context';
-import type { Literal, VariableRef } from '../types';
-import { lookupVariableRef } from './variable-ref';
+import type { Context } from '../format-context';
+import type { Literal, VariableRef } from './types';
+import { lookupVariableRef } from './resolve-variable.js';
 
 /** @internal */
 export function resolveValue(
