+++ packages/mf2-messageformat/src/messageformat.ts
@@ -1,10 +1,10 @@
 import { asDataModel, parseMessage } from './cst-parser/index.js';
-import { resolveExpression } from './data-model/expression/index.js';
-import { UnresolvedExpression } from './data-model/expression/variable-ref.js';
-import { formatMarkup } from './data-model/markup.js';
+import { formatMarkup } from './data-model/format-markup.js';
+import { resolveExpression } from './data-model/resolve-expression.js';
+import { UnresolvedExpression } from './data-model/resolve-variable.js';
 import type { Message } from './data-model/types.js';
 import { MessageError, MessageResolutionError } from './errors.js';
-import type { Context } from './format-context';
+import type { Context } from './format-context.js';
 import {
   MessageFunctions,
   MessagePart,
